+++ src/d_deh.h
@@ -489,7 +489,6 @@ extern char         *s_CAPTION_FREEDM;
 extern char         *s_CAPTION_BTSXE1;
 extern char         *s_CAPTION_BTSXE2;
 extern char         *s_CAPTION_BTSXE3;
-extern char         *s_CAPTION_BTSXE3PR;
 extern char         *s_CAPTION_REKKR;
 extern char         *s_CAPTION_REKKRSL;
 extern char         *s_CAPTION_ANOMALYREPORT;