+++ src/d_main.c
@@ -934,7 +934,7 @@ void D_CheckSupportedPWAD(char *filename)
     else if (M_StringCompare(leaf, "btsx_e3b.wad"))
         BTSX = BTSXE3 = BTSXE3B = true;
     else if (M_StringCompare(leaf, "btsxe3pr.wad"))
-        BTSX = BTSXE3PR = true;
+        BTSX = BTSXE3 = true;
     else if (M_StringCompare(leaf, "e1m4b.wad"))
         E1M4B = true;
     else if (M_StringCompare(leaf, "e1m8b.wad"))